Refactor minivan config to not use a layer hack
There is backlight support built into TMK core, so rather than injecting
a hacky function on layer switching for LED toggling, you can instead
add in support for the existing system and using the `ACTION_BACKLIGHT_`
commands.  For now, this toggles between off and 100% on.  A software
PWM can be done (look at the sleep led code) to do breathing, various
intensities, etc.

Also cleaned up the minivan codebase (removed incorrect/left over
keymaps for other GH60 layouts, fixed the README, added in comment
sections for the keymap layouts to be visualized, etc).  This does not
have updated keymaps for the arrow layout (but should not be much work)
or the additional layouts that Evan has been working on.